title bar
only show progress bar in title bar when in a Race session. Updated the session bar to not show the "m" in minutes for the time container if there's no elapsed or remaining time. Enabled ability to hide/show standings and relatives separate from each other when on track only.