fix for multi-class highlight relatives
In a multi-class race, the multiclass var was being set by only the drivers visible in the relatives box. So if a single class was in the relatives, the highlight color was used instead of the class color. Updated to utilize same functionality as standings so it looks at all drivers registered instead of just those who are in the few rows of the relatives.