fix: respect min class drivers in standings
In standings, the minimum number of drivers in class wasn't being respected. For example if you are p40 of 40, and have 3 for buffer, 10 for min class to display, and 1 for leaders of class to show, it would show p1, p37, p38, p39, p40. It would only show the buffer of 3 of cars ahead of you, and the one leader. This fix makes it so that even if you are last with a low buffer, it will show drivers ahead of you to make sure you reach your 10 min.